client day 17

+++ controllers/users.js
@@ -28,12 +28,22 @@ router.post("/add", async (req, res) => {
       email: req.body.email,
       phoneNumber: req.body.phoneNumber,
       password: await bcrypt.hash(req.body.password, 10),
-      type: req.body.type,
       createdOn: new Date()
     }
 
-    if(req.body.type === userTypes.USER_TYPE_STANDARD)
+    if(req.user.type === userTypes.USER_TYPE_STANDARD)
+    {
+      record.departmentId = req.user.departmentId;
+      record.type = userTypes.USER_TYPE_STANDARD;
+    }else
+    {
+      record.type = req.body.type;
+      if(req.body.type === userTypes.USER_TYPE_STANDARD)
       record.departmentId = req.body.departmentId;
+
+    }
+
+   
     const user = new User(record)
 
     await user.save();
@@ -60,6 +70,9 @@ router.post("/edit", async (req, res) => {
     const user = await User.findById(req.body.id);
     if (!user) throw new Error("User does not exists");
 
+    if(req.user.type === userTypes.USER_TYPE_STANDARD && user.departmentId.toString() !== req.user.departmentId.toString())
+      throw new Error('Invalid Request');
+
 
     const record = {
       name: req.body.name,
@@ -85,7 +98,7 @@ router.post("/edit", async (req, res) => {
 });
 
 
-router.delete("/delete", async (req, res) => {
+router.post("/delete", async (req, res) => {
   try {
     if (!req.body.id) throw new Error("User id is required");
     if (!mongoose.isValidObjectId(req.body.id))
@@ -93,6 +106,13 @@ router.delete("/delete", async (req, res) => {
 
     const user = await User.findById(req.body.id);
     if (!user) throw new Error("User does not exists");
+
+    if(req.body.id === req.user._id.toString())
+      throw new Error('Invalid Request')
+    
+    if(req.user.type === userTypes.USER_TYPE_STANDARD && user.departmentId.toString() !== req.user.departmentId.toString())
+      throw new Error('Invalid Request');
+    
 
     await User.findByIdAndDelete(req.body.id);
